+++ src/SolidBody.h
@@ -26,13 +26,13 @@ public:
 	double rho;     // density
 	double V;       // volume
 	std::vector<Node> Nodes;   // Nodes of the SolidBody mesh
-	int Nn;                    // Number of Nodes
+	size_t Nn;                    // Number of Nodes
 	
 	SolidBody(double x, double y, double ux, double uy, double omega, double rho, int Nn, bool moving);
 	~SolidBody();
 	void velocities();      // calculates the velocities in all Nodes of the SolidBody
 	void move(double d_t); // move Solid using $uc$ and $omega$
-	double ds(int i);
+	double ds(size_t i);
 };
 
 class Circle : public SolidBody{
